requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-anthropic==0.3.0
langchain-openai==0.2.10
anthropic==0.39.0
openai==1.55.0
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.0
pydantic==2.10.3
pydantic-settings==2.6.1
requests==2.32.3
beautifulsoup4==4.12.3
newspaper3k==0.2.8
lxml==5.1.0
python-docx==1.1.2
markdown2==2.5.1
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
tenacity==9.0.0
httpx==0.27.2
